Move `RankingsScope` to own file and fix mapper strings
